% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconv_wrapper.R
\name{deconv_wrap}
\alias{deconv_wrap}
\title{This function is a wrapper for ensemble deconvolution}
\usage{
deconv_wrap(
  count_bulk,
  meta_bulk = NULL,
  ref_list,
  customed_markers = NULL,
  markers_range = NULL,
  true_frac = NULL,
  params = NULL,
  ncv_input = 2,
  outpath,
  parallel_comp = FALSE,
  ncore,
  os = "win",
  myseed = 2020,
  feature.select = FALSE,
  clustering = TRUE,
  rm.duplicated = FALSE,
  mrkpen = FALSE,
  dmeths = NULL,
  cur_mrk,
  trueMet
)
}
\arguments{
\item{trueMet}{character}

\item{allfold_res}{all gene deconvolution results}

\item{allgene_res}{two folds deconvolution results}

\item{df_true}{Optional true cell type fraction}
}
\description{
This function is a wrapper for ensemble deconvolution
}
