% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsDeconv.R
\name{EnsDeconv}
\alias{EnsDeconv}
\title{This function is a wrapper for EnsDeconv}
\usage{
EnsDeconv(
  count_bulk,
  ref_list,
  customed_markers = NULL,
  true_frac = NULL,
  params = NULL,
  outpath = NULL,
  parallel_comp = FALSE,
  ncore,
  rm.duplicated = FALSE,
  mrkpen = FALSE,
  markers_range = NULL,
  dmeths = NULL,
  inrshiny = FALSE
)
}
\arguments{
\item{count_bulk}{Bulk gene expression.

(Required)  Two-dimensional numeric. Must be in gene x sample format. Must implemented \code{as.matrix}.
(Optional) In original scale.}

\item{ref_list}{List of of list. Must implement as.list.

(Required) The i-th element of the top-level list is a list of \code{ref_matrix}, \code{meta_ref}. Names of the top level list should be vector of
\code{data_name}, namely : bulk-reference.

\itemize{
\item{'ref_matrix'}{ Reference matrix, rows are genes, columns are samples.}
\item{'meta_ref'}{ Meta data for reference data. Must include two variables: "SamplesName" (column names of ref_matrix), "deconv_clust" (deconvolution cluster, eg. cell types).}
\item{'data_name'}{Data description. Character in format "bulk data name_reference data name"}
}}

\item{customed_markers}{(Optional) Self-defined markers.}

\item{true_frac}{(Optional) True cell type proportions for bulk gene expression
Two-dimensional numeric. Must be in samples by cell type.}

\item{params}{Parameters Dataframe for ensemble learning, more details could refer to \code{get_params}.}

\item{outpath}{(Optional) Path to save output.}

\item{parallel_comp}{Logical. Use parallel computing or not. Default: FALSE.}

\item{ncore}{The number of cores to use for parallel execution.}

\item{rm.duplicated}{Logical. Remove duplicated genes after maker gene selection. Default: FALSE.}

\item{mrkpen}{Logical. Apply markerpen on marker gene list. Default: FALSE.}

\item{markers_range}{(Optional) Specific for markerpen.}
}
\value{
A list containing the output of the EnsDeconv algorithm (EnsDeconv) and a list output from each scenario (allgene_res).
}
\description{
This function is a wrapper for EnsDeconv
}
