% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_all_res.R
\name{gen_all_res}
\alias{gen_all_res}
\title{A parent Function}
\usage{
gen_all_res(
  count_bulk,
  meta_bulk = NULL,
  ref_matrix,
  meta_ref = NULL,
  customed_markers = NULL,
  markers_range = NULL,
  true_frac = NULL,
  params = NULL,
  ncv_input = 2,
  outpath,
  data_name = NULL,
  parallel_comp = FALSE,
  ncore = NULL,
  os = "win",
  myseed = 2020,
  feature.select = FALSE,
  clustering = TRUE,
  rm.duplicated = TRUE,
  mrkpen = FALSE,
  dmeths = NULL
)
}
\arguments{
\item{count_bulk}{Bulk gene expression.

(Required)  Two-dimensional numeric. Must be in gene x sample format. Must implemented \code{as.matrix}

(Optional) In original scale.}

\item{meta_bulk}{Meta data for bulk gene expression.

(Optional) Dataframe, much include character variable SamplesName.

(Required) If reference data is single cell rna seq.}

\item{ref_matrix}{Reference matrix.

(Required) Two-dimensional numeric.  Could be signature matrix (gene x celltype), non-single cell RNA seq reference matrix (gene x sample),
single cell RNA seq reference matrix (gene x cell). Could be  \code{as(ref_matrix, "dgCMatrix")} or \code{as.matrix}}

\item{meta_ref}{Meta data for reference matrix.

(Optional) Dataframe, much include character variable SamplesName, and deconv_clust.

(Required)  If reference data is single cell rna seq.

deconv_clust : character strings for "cell types".}

\item{customed_markers}{Self-defined markers.

(Optional) List of one-dimensional string Names of the list should match the deconv_clust.}

\item{markers_range}{Specific for markerpen.

(Optional)}

\item{true_frac}{True cell type proportions for bulk gene expresseeion.

(Optional) Two-dimensional numeric. Must be in samples by celltype.}

\item{params}{Parameters dataframe for ensemble learning, more details could refer to \code{get_params}.}

\item{ncv_input}{Number of folds.

(Optional) Default is 2}

\item{outpath}{(Required) Path to save output.}

\item{data_name}{Data description.

(Optional) Only input when you want default params. Character in format "Bulk data name_reference data name"}

\item{parallel_comp}{Logical.}

\item{ncore}{The number of cores to use for parallel execution.}

\item{os}{Operation system. Default is "win". For mac user, please specify as "OS"}
}
\description{
This function allows you generate results for cross validation using different deconvolution methods.
}
\examples{

data(testdata)

params  = get_params("none","none","log","singlecell-rna","test_test",20,"p.value")

gen_all_res(count_bulk =testdata$count_bulk ,meta_bulk = testdata$meta_bulk,ref_matrix = testdata$ref_list$Nowakowski$ref_matrix,
meta_ref =  testdata$ref_list$Nowakowski$meta_ref,true_frac = testdata$true_frac, params = params, ncv_input =1,
outpath = "D:/ensemble deconvolution/test/",data_name = "test-test",parallel_comp = TRUE, ncore = 2)
}
