% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dens_all_mae.R
\name{gen_dens_all_mae}
\alias{gen_dens_all_mae}
\title{This function is used to generate density plot of pearson correlation and MAE}
\usage{
gen_dens_all_mae(
  re_org_out_all,
  tit,
  customed_mrk = FALSE,
  cust,
  notation = T,
  note = NULL
)
}
\description{
This function is used to generate density plot of pearson correlation and MAE
}
