% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric_wrap.R
\name{get_metric_wrap}
\alias{get_metric_wrap}
\title{This function is used for generate metric}
\usage{
get_metric_wrap(res_list, true, trueMet, cal_mean = FALSE)
}
\arguments{
\item{res_list}{: list, each element has a, p, and ensemble}

\item{true}{: true fraction}

\item{trueMet}{: Bulk data name}

\item{cal_mean}{: logical, calculate mean}

\item{cmap}{: named cell type}
}
\description{
This function is used for generate metric
}
