% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{This function generates parameters/settings for ensemble deconvolution}
\usage{
get_params(
  TNormalization = c("CPM", "none", "TPM", "TMM", "QN"),
  CNormalization = c("CPM", "none", "TPM", "TMM", "QN"),
  Scale = c("log", "linear"),
  data_type,
  data_name,
  n_markers = 50,
  Marker.Method = c("t", "wilcox", "combined", "none", "p.value", "regression"),
  dmeths = NULL,
  teqc = TRUE,
  batchcorrec = FALSE,
  RB_only = TRUE
)
}
\arguments{
\item{Scale}{Scaling methods for gene expression data.

(Option) Vector of string. Defaults are "log" and "linear"}

\item{data_type}{Reference data data type

(Required) One-dimensional string. Options are

\itemize{
\item{'singlecell-rna'}{ Single cell RNA seq reference.}
\item{'microarray'}{ Microarray reference.} }}

\item{data_name}{Name for bulk_reference

(Options) One-dimensional string. Can be in format : bulk_reference}

\item{n_markers}{Number of markers

(Option) One-dimensional numeric, default is 150

How many markers genes to use for deconvolution. Can either be a single integer, vector of integers. All cell types use the same number of markers.}

\item{dmeths}{Deconvolution methods to used. Default is NULL, EnsDeconv will run all possible deconvolution methods for specific reference data sets.}

\item{Marker.Method}{Method used to choose marker genes

\itemize{
\item{'ratio'}{ selects and ranks markers by the ratio of the mean expression of each gene in each cell type to the mean of that gene in all other cell types.}
\item{'regression '}{ selects and ranks markers by estimated regression coefficients in a series of regressions with single covariate that is indicator of
each type.}
\item{'diff'}{ selects and ranks markers based upon the difference, for each cell type, between the median expression of a gene by each cell type and the
median expression of that gene by the second most highly expressed cell type.}
\item{'p.value'}{ selects and ranks markers based upon the p-value of a t-test between the median expression of a gene by each cell type and the median
expression of that gene by the second most highly expressed cell type.}
\item{'t'}{ Perform pairwise Welch t-tests between groups of cells, possibly after blocking on uninteresting factors of variation.}
\item{'wilcox'}{ Perform pairwise Wilcoxon rank sum tests between groups of cells, possibly after blocking on uninteresting factors of variation.}
\item{'binom'}{ Perform pairwise binomial tests between groups of cells, possibly after blocking on uninteresting factors of variation.}
\item{'TOAST'}{ Iteratively searches for cell type-specific features }
\item{'Hedge'}{ Iteratively searches for cell type-specific features }
\item{'roc'}{ Iteratively searches for cell type-specific features }
}}

\item{teqc}{Logical. Use same normalization between bulk data and reference data. Default: TRUE.}

\item{batchcorrec}{Logical. Apply batch correction or not. Default: FALSE}

\item{Normalization}{normalization methods for bulk gene expression and reference gene expression

(Option) Vector of string. Defaults are "CPM", "TPM" and "none"}
}
\value{
a data.frame that each row corresponds to a specific deconvolution scenario/setting
}
\description{
This function generates parameters/settings for ensemble deconvolution.
}
