% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_wrap.R
\name{get_res_wrap}
\alias{get_res_wrap}
\title{This function is used to get ensemble results}
\usage{
get_res_wrap(
  allfold_res,
  allgene_res,
  df_true,
  trueMet,
  feature_selection = FALSE,
  lambda = NULL,
  incl_lasso = FALSE,
  get_metric = TRUE,
  r2 = FALSE,
  r2_ind = NULL,
  scale = "log",
  transformation = "none",
  optimization = TRUE
)
}
\arguments{
\item{allfold_res}{all gene deconvolution results}

\item{allgene_res}{two folds deconvolution results}

\item{df_true}{Optional true cell type fraction}

\item{trueMet}{character}
}
\description{
This function is used to get ensemble results
}
