% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare data for analysis}
\usage{
prepare(dat, bulk_tissue)
}
\arguments{
\item{dat}{A Seurat object containing single-cell RNA data.}

\item{bulk_tissue}{A matrix containing bulk RNA sequencing data.}
}
\value{
A list containing processed bulk and single-cell data.
}
\description{
This function prepares single-cell and bulk RNA sequencing data for further analysis.
}
