% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_all.R
\name{process_all}
\alias{process_all}
\title{This function is used for process output}
\usage{
process_all(
  allfold_res,
  allgene_res,
  df_true = NULL,
  trueMet = NULL,
  feature_selection = TRUE,
  lambda = NULL,
  incl_lasso = FALSE,
  get_metric = TRUE,
  rm_dmeths = TRUE,
  r2 = FALSE,
  r2_ind = NULL,
  scale = "log",
  transformation = "CPM",
  optimization = FALSE
)
}
\arguments{
\item{allfold_res}{List of output from gen_all_res or gen_all_res_list of two fold

Each element of list is result for one scenario, which contains element \code{a}, \code{p}, and \code{ensemble}.

\itemize{
\item{'a'}{  Contains p_hat (list, deconvolution results of each fold), p_truth (list, true cell type proportion), n (list, number of markers of each fold), time, markers (list, markers gene of each fold), all_time (computation time)}
\item{'p'}{  Parameters}
\item{'ensemble'}{  Testing data for each fold}
}}

\item{allgene_res}{List of output from gen_all_res or gen_all_res_list of one fold

Each element of list is result for one scenario, which contains element \code{a}, \code{p}, and \code{ensemble}.

\itemize{
\item{'a'}{  Contains p_hat (list of deconvolution results), p_truth (list, true cell type proportion), n (list, number of markers of each fold), time, markers (list, markers gene of each fold), all_time (computation time)}
\item{'p'}{  Parameters}
\item{'ensemble'}{  Testing data for}
}}

\item{df_true}{Dataframe, True cell type expression, samples*cell type}

\item{trueMet}{Character, name of bulk data}

\item{feature_selection}{Logical, conducting feature selection or not}

\item{lambda}{(Optional)  tuning parameter for feature selection}

\item{get_metric}{Logical, calculating metric or not}

\item{rm_dmeths}{Logical, remove several deconvolution methods or not when ensemble}
}
\description{
This function is used for process output
}
