% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsDeconv.R
\name{EnsDeconv}
\alias{EnsDeconv}
\title{EnsDeconv: A Wrapper Function for Ensemble Deconvolution}
\usage{
EnsDeconv(
  count_bulk,
  ref_list,
  enableFileSaving = FALSE,
  outpath = NULL,
  parallel_comp = FALSE,
  ncore = 5,
  true_frac = NULL,
  params = NULL,
  inrshiny = FALSE
)
}
\arguments{
\item{count_bulk}{Bulk gene expression data.
The count_bulk parameter expects a two-dimensional numeric matrix in a gene-by-sample format. 
        It must be convertible using \code{as.matrix}. Optionally, the data can be in its original scale.}

\item{ref_list}{Reference data list.
The ref_list is a list of lists, where each sublist contains \code{ref_matrix} and \code{meta_ref}. 
        The top-level list should be named with a vector of \code{data_name}, indicating the bulk-reference pair.
        The sublists should contain:
        \itemize{
        \item{ref_matrix}{A matrix with rows as genes and columns as samples.}
        \item{meta_ref}{Metadata for reference data, including "SamplesName" (column names of ref_matrix) 
          and "deconv_clust" (deconvolution clusters, e.g., cell types).}
        \item{data_name}{A description of the data, formatted as "bulk data name_reference data name".}
        }}

\item{enableFileSaving}{Enable Saving of Intermediate Output
(Optional) A boolean flag that controls the saving of intermediate outputs as separate files. 
        When set to TRUE, intermediate outputs of the analysis will be saved to files. 
        If not explicitly set, this parameter defaults to FALSE, meaning that intermediate 
        outputs will not be saved by default.}

\item{parallel_comp}{Use parallel computing.
(Optional) A logical flag indicating whether to perform computations in parallel. 
        Defaults to FALSE.}

\item{ncore}{Number of cores for parallel execution.
(Optional) Sets the number of cores for parallel processing when \code{parallel_comp} is TRUE. 
        Default is 5. Only effective if parallel computing is enabled.}

\item{true_frac}{True cell type proportions.
(Optional) A two-dimensional numeric matrix indicating the true cell type proportions 
        in the samples. The matrix should be formatted with samples as rows and cell types as columns.}

\item{params}{Ensemble learning parameters.
(Optional) A dataframe specifying parameters for ensemble learning. 
        For more details, refer to the \code{get_params} function.}

\item{outputPath}{Destination for Saved Output Files
(Optional) Specifies the ile path where output files should be saved, applicable 
        only if \code{enableFileSaving} is set to TRUE. Providing this path directs the function 
        to save all intermediate output files to the specified location. 
        If \code{enableFileSaving} is FALSE or not set, the value of "outputPath" is ignored.
        This parameter should be a valid file system path.}
}
\value{
A list containing two elements:
        - \code{EnsDeconv}: The output of the EnsDeconv algorithm.
        - \code{allgene_res}: A list of results from each scenario analyzed.
}
\description{
This function serves as a wrapper for the EnsDeconv algorithm, providing an interface for ensemble deconvolution analysis on gene expression data.
}
\examples{
# Example usage
# EnsDeconv(count_bulk, ref_list)

}
