% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_all_res_list.R
\name{gen_all_res_list}
\alias{gen_all_res_list}
\title{This function is used when there are multiple references.}
\usage{
gen_all_res_list(
  count_bulk,
  meta_bulk = NULL,
  ref_list,
  customed_markers = NULL,
  markers_range = NULL,
  true_frac = NULL,
  params = NULL,
  outpath,
  parallel_comp = FALSE,
  ncore,
  os = "win",
  rm.duplicated = FALSE,
  mrkpen = FALSE,
  dmeths = NULL
)
}
\arguments{
\item{count_bulk}{Bulk gene expression.
(Required)  Two-dimensional numeric. Must be in gene x sample format. Must implemented \code{as.matrix}
(Optional) In original scale.}

\item{meta_bulk}{Meta data for bulk gene expression.
(Optional) Dataframe, much include character variable SamplesName.
(Required) If reference data is single cell rna seq.}

\item{ref_list}{List of of list. Must implement as.list.


(Required) The i-th element of the top-level list is a list of \code{ref_matrix}, \code{meta_ref}. Names of the top level list should be vector of
\code{data_name}, namely : bulk-reference.

\itemize{
\item{'ref_matrix'}{ Reference matrix.}
\item{'meta_ref'}{ Meta data for reference matrix..}
\item{'data_name'}{Data description. Character in format "Bulk data name_reference data name"}
}}

\item{customed_markers}{Self-defined markers.
(Optional) List of one-dimensional string Names of the list should match the deconv_clust.}

\item{markers_range}{Specific for markerpen.
(Optional)}

\item{true_frac}{True cell type proportions for bulk gene expresseeion.
(Optional) Two-dimensional numeric. Must be in samples by celltype.}

\item{params}{Parameters dataframe for ensemble learning, more details could refer to \code{get_params}.}

\item{outpath}{(Optional) Path to save output.}

\item{parallel_comp}{Logical.}

\item{ncore}{The number of cores to use for parallel execution.}

\item{os}{Operation system. Default is "win". For mac user, please specify as "OS"}

\item{rm.duplicated}{Logical. Remove duplicated genes after maker gene selection. Default: FALSE.}

\item{mrkpen}{Logical. Apply markerpen on marker gene list. Default: FALSE.}

\item{data_name}{Data description.
(Optional) Only input when you want default params. Character in format "Bulk data name-reference data name"}
}
\description{
This function is used when there are multiple references.
}
