% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{A data example}
\value{
A list containing
\item{count_bulk}{}
\item{meta_bulk}{}
\item{ref_list}{}
\item{true_frac}{}
}
\description{
A data list for demonstration.
}
\examples{

data(testdata)

}
