% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{testdata.}
\format{
A list of count_bulk and ref_list:
\describe{
  \item{count_bulk}{sample bulk data, where rows are genes and columns are samples}
  \item{ref_list}{list of references}
  ...
}
}
\usage{
testdata
}
\description{
A dataset containing a sample bulk dataset and sample reference data.
}
\keyword{datasets}
